fix(storage): heal path-hash indexes and re-arm data_sync on index readiness
Decouple index repair from data_sync liveness so residual Entropy holes
recover without waiting for a clean global heal or an epoch boundary.

Index heal (startup, assignment, and periodic while needs_retry):
- Scan path-hash gaps, re-migrate covering blocks in 128-block passes
- Soft-skip missing history/timeouts; emit unrepaired metrics
- Bail mid-migrate on shutdown via Shutdown clone on the service
- Delete path-hash keys on clear so reorg clears are real, healable gaps

Data sync:
- Drop heal-gated SyncPartitions unblock flag
- Re-queue Blocked(MissingDataRootIndex) when local data_root_and_tx_offset_at
  succeeds, capped by free pending slots and a probe budget

Also extract storage_module_service into mod + index_heal for maintainability.
